requirements.txt
annotated-types==0.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
dataclass-wizard==0.35.0
idna==3.10
python-dotenv==1.0.0
requests==2.32.4
todoist_api_python==3.1.0
typing_extensions==4.14.0
urllib3==2.5.0
